requirements.txt
streamlit==1.28.1
google-generativeai==0.3.2
PyPDF2==3.0.1
reportlab==4.0.4
python-dotenv==1.0.0
pandas==2.1.3
numpy==1.24.3
